requirements.txt
MouseInfo==0.1.3
mss==6.1.0
numpy==1.22.4
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.1.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
PyRect==0.2.0
PyScreeze==0.1.28
pytesseract==0.3.9
pytweening==1.0.4
screeninfo==0.8